+++ assets/src/scripts/d3-rendering/markers.spec.js
@@ -1,5 +1,6 @@
 import {select} from 'd3-selection';
-import {lineMarker, circleMarker, rectangleMarker, textOnlyMarker} from 'd3render/markers';
+
+import {lineMarker, circleMarker, rectangleMarker, textOnlyMarker} from './markers';
 
 describe('Markers module', () => {